Recommender System
Book Rental Recommendation System
Bookrent is the largest online and offline book rental chain in India. The company charges a fixed rental fee for a book per month.  Lately, the company has been loosing its user base. The main reason for this is that users are not able to choose the right books for themselves. The company wants to solve this problem and increase its revenue and profits. Different students will be required to work on different components of the broad problem statement and the relavant dataset will be made available to students.

Tasks performed:

1 Read the books dataset and explore it.

2 Clean up NaN values

3 Read the data where ratings are given by the users.

4 Take a quick look at the number of unique users and books.

5 Perform data conversion for consistency.

6 Split your data into two sets (Training and testing)

7 Calculate the Similarity.